test(checkbox): migrate item
