+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-inset-list-ios-ltr-Mobile-Firefox-linux.png
