+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-inset-list-md-ltr-Mobile-Safari-linux.png
