+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-item-color-ios-ltr-Mobile-Firefox-linux.png
