+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-item-color-md-ltr-Mobile-Chrome-linux.png
