+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-list-ios-ltr-Mobile-Safari-linux.png
