+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-list-ios-rtl-Mobile-Safari-linux.png
