+++ core/src/components/checkbox/test/item/checkbox.e2e.ts-snapshots/checkbox-list-md-ltr-Mobile-Safari-linux.png
